Rename args (if funcrefs), tweak self-ref checks
PrettyPrint and ShallowPrint functions may be given Funcrefs;
when this happens, the identifiers for those Funcrefs must start with
capital letters to prevent vim from throwing errors.

Add, and attempt to pass, test cases for printing Funcrefs and Partials.

Check for "immediate child is self-reference" in PrettyPrintList and
PrettyPrintDict. This better handles a few edge cases related to Partial
printing, and prevents us from needing to initialize the "seen_objs"
list in function(typevim#PrettyPrint).

This also causes us to fail an older test case; need to investigate
further.