Don't "blindly" propagate Resolve, Reject calls
Only propagate a Rejection past an error handler if the error handler
itself threw an exception; or, if a Resolve threw an exception, reject
the next Promise in the chain.

Add more test cases for this behavior. Start trying to comply with the
Promise/A+ spec, though this might not be entirely possible in VimL.