Disregard 'liveness' w/ unhandled rejections
It is *not* inevitable that a *successfully handled* Promise rejection
will "erroneously" reject the "dead" Promises returned by a call to
`Then()`. If there is an error handler in the chain, and it returns a
value without throwing an exception, the next Promise in the chain will
be resolved.

Then, it is sufficient to check whether we're at a "leaf" node, then
check whether an error handler exists, and throw if none could be found.