Don't give l:self when executing Promise callbacks
Store given callbacks in a member list inside of HandlerAttachment,
rather than storing them directly inside of HandlerAttachment. This
avoids "binding" the HandlerAttachment with the callback, if the
callback is a dict function without an already-bound dictionary.

This is for compliance with 2.2.5 of Promises/A+, which essentially
requires that `this` be undefined when invoking a provided callback
function, unless that function is somehow "bound" to a predetermined
`this` before/during the call to `then()`.