search() arg parsing helper, write chunk fetcher
In preparation for writing a faster FindLines() function, that fetches
lines from the buffer by chunks and tests a {regexp} against each one.
This will hopefully be faster than Buffer.search(), since it won't have
to open the buffer in a new tab.

Write tests for the above, including basic tests of the more subtle
search() flags.